+++ NEWS.md
@@ -4,6 +4,17 @@ This page describes the most important changes in `TypeUtils`. The format is bas
 [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 [Semantic Versioning](https://semver.org).
 
+## Unreleased
+
+### Added
+
+- `B = lazymap([T::Type,], f, A[, f_inv])` yields an object `B` that lazily maps function
+  `f` to array or iterator `A`. Optional argument `T` is the element-type of `B` which ia
+  automatically inferred if not specified. If `A` is an (abstract) array, `f_inv` my be
+  provided to specify the inverse of `f`. Related type alias `LazyMap{T,N,F,typeof(A)}` can
+  be used to specialize on the type `B` with `N = ndims(A)` and `F = typeof(f)`.
+
+
 ## Version 1.13.0 (2025-12-19)
 
 ### Added